multi agent/requirements.txt
# Core dependencies
PyYAML>=6.0
PyQt5>=5.15
docker>=7.0.0

google-generativeai
python-dotenv
# LLM service dependencies
openai>=1.0.0
google-generativeai>=0.3.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
# Core requirements
python-dotenv>=1.0.0
requests>=2.31.0

# LLM SDKs
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Optional framework integrations
langchain>=0.1.0
fastapi>=0.95.2
uvicorn>=0.22.0

# Development tools
black>=23.7.0
flake8>=6.0.0
pytest>=7.4.0